requirements.txt
uvicorn[standard]
gunicorn
pydantic
pony
psycopg2
python-dotenv
pydantic[dotenv]
fastapi
python-multipart
python-jose[cryptography]
passlib[bcrypt]
decouple
python-decouple
haversine
xmltodict
requests
pandas
nltk
sklearn
fastapi-cache